Trekk ut regelsett meny og view i egne komponenter. Legg til useEffect som oppdaterer aktivtregelsett når behandling oppdateres.
