+++ app/components/behandling/Behandling.module.css
@@ -2,60 +2,3 @@
   display: grid;
   grid-template-columns: 350px 1fr;
 }
-
-.regelsettHeading {
-  padding: var(--a-spacing-2);
-}
-
-.regelsettMeny {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  border-right: 1px solid var(--a-border-subtle);
-}
-
-.divider {
-  border-bottom: 1px solid var(--a-border-subtle);
-}
-
-.regelsettMeny li {
-  margin: 0;
-}
-
-.opplysningListe {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.akitvtRegelsett {
-  background-color: var(--a-bg-subtle);
-  box-shadow: 4px 0 0 0 inset var(--a-border-action-selected);
-}
-
-.regelsettMeny li button {
-  width: 100%;
-  height: 100%;
-  border: none;
-
-  display: flex;
-  align-items: center;
-
-  padding: var(--a-spacing-2);
-  background-color: inherit;
-  text-align: left;
-  cursor: pointer;
-}
-
-.regelsettMeny li button:hover {
-  background-color: var(--a-surface-hover);
-}
-
-.regelsettMeny li button svg {
-  margin-right: var(--a-spacing-2);
-}
-
-.circleIcon {
-  margin-left: var(--a-spacing-2);
-  color: var(--a-surface-warning-moderate);
-}
