+++ app/components/regelsett/Regelsett.module.css
@@ -1,0 +1,9 @@
+.regelsettHeading {
+  padding: var(--a-spacing-2);
+}
+
+.opplysningListe {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
