+++ app/routes/oppgave.$oppgaveId.behandle.tsx
@@ -1,7 +1,7 @@
 import { DocPencilIcon, TasklistSendIcon } from "@navikt/aksel-icons";
 import { Tabs } from "@navikt/ds-react";
 import type { ActionFunctionArgs } from "@remix-run/node";
-import { Outlet, useActionData, useLocation } from "@remix-run/react";
+import { Outlet, useActionData } from "@remix-run/react";
 import { useState } from "react";
 
 import { Behandling } from "~/components/behandling/Behandling";
@@ -17,7 +17,6 @@ export async function action({ request, params }: ActionFunctionArgs) {
 }
 
 export default function Oppgave() {
-  const { key } = useLocation();
   const actionData = useActionData<typeof action>();
   const [aktivTab, setAktivTab] = useState("behandling");
   useHandleAlertMessages(isAlert(actionData) ? actionData : undefined);
@@ -37,7 +36,7 @@ export default function Oppgave() {
           </Tabs.List>
 
           <Tabs.Panel value="behandling">
-            <Behandling key={key} />
+            <Behandling />
           </Tabs.Panel>
 
           <Tabs.Panel value="melding-om-vedtak">
